Implement `INCLUDE_DRBB` environment variable
