+++ .env-sample
@@ -7,3 +7,5 @@ export SET SHEP_API=http://[fqdn shep api]/api/v0.1
 export SET HOLD_REQUEST_NOTIFICATION="notification"
 export SET CLOSED_LOCATIONS="all;Library for the Performing Arts"
 export SET SEARCH_RESULTS_NOTIFICATION="notification"
+
+export SET INCLUDE_DRBB=[boolean]
