+++ src/app/components/Application/Application.jsx
@@ -33,7 +33,10 @@ class Application extends React.Component {
   }
 
   getChildContext() {
-    return { media: this.state.media };
+    return {
+      media: this.state.media,
+      includeDrbb: appConfig.includeDrbb,
+    };
   }
 
   componentDidMount() {
@@ -152,6 +155,7 @@ Application.contextTypes = {
 
 Application.childContextTypes = {
   media: PropTypes.string,
+  includeDrbb: PropTypes.bool,
 };
 
 export default Application;
