+++ src/app/components/ResultsList/ResultsList.jsx
@@ -97,6 +97,7 @@ class ResultsList extends React.Component {
   render() {
     const results = this.props.results;
     let resultsElm = null;
+    const { includeDrbb } = this.context;
 
     if (!results || !_isArray(results) || !results.length) {
       return null;
@@ -107,7 +108,9 @@ class ResultsList extends React.Component {
     return (
       <ul
         id="nypl-results-list"
-        className={`nypl-results-list ${Store.getState().isLoading ? 'hide-results-list ' : 'drbb-integration'}`}
+        className={
+          `nypl-results-list${Store.getState().isLoading ? ' hide-results-list' : ''}${includeDrbb ? ' drbb-integration' : ''}`
+        }
       >
         {resultsElm}
       </ul>
@@ -122,6 +125,7 @@ ResultsList.propTypes = {
 
 ResultsList.contextTypes = {
   router: PropTypes.object,
+  includeDrbb: PropTypes.bool,
 };
 
 export default ResultsList;
