+++ src/app/components/SccContainer/SccContainer.jsx
@@ -7,35 +7,38 @@ import Notification from '../Notification/Notification';
 
 import Store from '../../stores/Store';
 
-const SccContainer = props => (
-  <main className="main-page">
-    <LoadingLayer
-      status={Store.state.isLoading}
-      title={props.loadingLayerText}
-    />
-    <div className="nypl-page-header">
-      <div className="nypl-full-width-wrapper drbb-integration filter-page">
-        <div className="nypl-row">
-          <div className="nypl-column-full">
-            <Breadcrumbs type={props.breadcrumbsType} />
-            <Notification notificationType="searchResultsNotification" />
-            <h1
-              aria-label={props.bannerOptions.ariaLabel || props.bannerOptions.text}
-            >
-              { props.bannerOptions.text }
-            </h1>
-            { props.extraBannerElement }
+const SccContainer = (props, context) => {
+  const { includeDrbb } = context;
+  return (
+    <main className="main-page">
+      <LoadingLayer
+        status={Store.getState().isLoading}
+        title={props.loadingLayerText}
+      />
+      <div className="nypl-page-header">
+        <div className={`nypl-full-width-wrapper filter-page${includeDrbb ? ' drbb-integration' : ''}`}>
+          <div className="nypl-row">
+            <div className="nypl-column-full">
+              <Breadcrumbs type={props.breadcrumbsType} />
+              <Notification notificationType="searchResultsNotification" />
+              <h1
+                aria-label={props.bannerOptions.ariaLabel || props.bannerOptions.text}
+              >
+                { props.bannerOptions.text }
+              </h1>
+              { props.extraBannerElement }
+            </div>
           </div>
         </div>
+        { props.secondaryExtraBannerElement }
       </div>
-      { props.secondaryExtraBannerElement }
-    </div>
-    { props.extraRow }
-    <div className="nypl-full-width-wrapper drbb-integration">
-      { props.mainContent }
-    </div>
-  </main>
-);
+      { props.extraRow }
+      <div className={`nypl-full-width-wrapper${includeDrbb ? ' drbb-integration' : ''}`}>
+        { props.mainContent }
+      </div>
+    </main>
+  );
+};
 
 SccContainer.propTypes = {
   mainContent: PropTypes.element,
@@ -50,7 +53,12 @@ SccContainer.propTypes = {
 SccContainer.defaultProps = {
   mainContent: null,
   extraBannerElement: null,
-  loadingLayerText: "Loading",
+  loadingLayerText: 'Loading',
+};
+
+SccContainer.contextTypes = {
+  router: PropTypes.object,
+  includeDrbb: PropTypes.bool,
 };
 
 export default SccContainer;
