+++ src/app/pages/SearchResults.jsx
@@ -14,7 +14,7 @@ import {
   basicQuery,
 } from '../utils/utils';
 
-const SearchResults = (props) => {
+const SearchResults = (props, context) => {
   const {
     searchResults,
     searchKeywords,
@@ -22,9 +22,15 @@ const SearchResults = (props) => {
     filters,
     page,
     field,
-    location,
     sortBy,
   } = props;
+
+  const {
+    includeDrbb,
+    router,
+  } = context;
+
+  const { location } = router;
 
   const [dropdownOpen, toggleDropdown] = useState(false);
 
@@ -69,7 +75,7 @@ const SearchResults = (props) => {
         mainContent={<SearchResultsContainer {...props} />}
         bannerOptions={
           {
-            text: "Search Results",
+            text: 'Search Results',
             ariaLabel: headerLabel,
           }
         }
@@ -94,7 +100,7 @@ const SearchResults = (props) => {
             />
             {
               selectedFiltersAvailable &&
-              <div className="nypl-full-width-wrapper selected-filters drbb-integration">
+              <div className={`nypl-full-width-wrapper selected-filters${includeDrbb ? ' drbb-integration' : ''}`}>
                 <div className="nypl-row">
                   <div className="nypl-column-full">
                     <SelectedFilters
@@ -109,7 +115,7 @@ const SearchResults = (props) => {
         }
         extraRow={
           <div className="nypl-sorter-row">
-            <div className="nypl-full-width-wrapper drbb-integration">
+            <div className={`nypl-full-width-wrapper selected-filters${includeDrbb ? ' drbb-integration' : ''}`}>
               <div className="nypl-row">
                 <div className="nypl-column-full">
                   <ResultsCount
@@ -145,7 +151,6 @@ SearchResults.propTypes = {
   searchKeywords: PropTypes.string,
   selectedFilters: PropTypes.object,
   page: PropTypes.string,
-  location: PropTypes.object,
   filters: PropTypes.object,
   field: PropTypes.string,
   sortBy: PropTypes.string,
@@ -153,6 +158,11 @@ SearchResults.propTypes = {
 
 SearchResults.defaultProps = {
   page: '1',
+};
+
+SearchResults.contextTypes = {
+  includeDrbb: PropTypes.bool,
+  router: PropTypes.obj,
 };
 
 export default SearchResults;
