+++ test/fixtures/resultsBibs.js
@@ -1,0 +1,244 @@
+const resultsBibs = [
+  {
+    result: {
+      '@id': 'res:b17692265',
+      extent: ['xii, 267 p. : ill. '],
+      note: ['Includes bibliographical references (p. 244-258) and index.'],
+      createdYear: 2007,
+      title: ['Hamlet without Hamlet', 'Hamlet without Hamlet /'],
+      creatorLiteral: ['De Grazia, Margreta.'],
+      createdString: ['2007'],
+      idOwi: ['urn:owi:58292525'],
+      dateStartYear: 2007,
+      suppressed: false,
+      identifier: [
+        'urn:bnum:17692265',
+        'urn:oclc:255835683',
+        'urn:lcc:PR2807',
+        'urn:lccCoarse:PR2199-3195',
+      ],
+      materialType: [
+        {
+          '@id': 'resourcetypes:txt',
+          prefLabel: 'Text',
+        },
+      ],
+      titleDisplay: ['Hamlet without Hamlet / Margreta De Grazia.'],
+      uri: 'b17692265',
+      numItems: 2,
+      numAvailable: 0,
+      uris: [
+        'b17692265',
+        'b17692265-i21106088',
+        'b17692265-i21106086',
+      ],
+      placeOfPublication: ['Cambridge, UK ; New York :'],
+      publicationStatement: ['Cambridge, UK ; New York : Cambridge University Press,'],
+      issuance: [
+        {
+          '@id': 'urn:biblevel:m',
+          prefLabel: 'monograph/item',
+        },
+      ],
+      publisherLiteral: ['Cambridge University Press,'],
+      items: [
+        {
+          '@id': 'res:i21106088',
+          uri: 'i21106088',
+          identifier: ['urn:SierraNypl:21106088'],
+          idNyplSourceId: {
+            '@type': 'SierraNypl',
+            '@value': 21106088,
+          },
+        },
+        {
+          '@id': 'res:i21106086',
+          uri: 'i21106086',
+          identifier: ['urn:SierraNypl:21106086'],
+          idNyplSourceId: {
+            '@type': 'SierraNypl',
+            '@value': 21106086,
+          },
+        },
+      ],
+    },
+  },
+  {
+    result: {
+      '@id': 'res:b17678038',
+      extent: ['195 p. : chiefly ill. ;'],
+      note: ['Includes bibliographical references (p. 244-258) and index.'],
+      createdYear: 2007,
+      title: ['Hamlet', 'Hamlet /'],
+      creatorLiteral: ['Vieceli, Emma.', 'Shakespeare, William, 1564-1616.'],
+      createdString: ['2007'],
+      idOwi: ['urn:owi:1910969040'],
+      dateStartYear: 2007,
+      suppressed: false,
+      identifier: [
+        'urn:bnum:17678038',
+        'urn:oclc:690493652',
+        'urn:lcc:PN6737.A66',
+        'urn:lccCoarse:PN6700-6790',
+      ],
+      materialType: [
+        {
+          '@id': 'resourcetypes:txt',
+          prefLabel: 'Text',
+        },
+      ],
+      titleDisplay: ['Hamlet without Hamlet / Margreta De Grazia.'],
+      uri: 'b17678038',
+      numItems: 2,
+      numAvailable: 0,
+      uris: [
+        'b17692265',
+        'b17692265-i21106088',
+        'b17692265-i21106086',
+      ],
+      placeOfPublication: ['New York :'],
+      publicationStatement: ['New York :'],
+      issuance: [
+        {
+          '@id': 'urn:biblevel:m',
+          prefLabel: 'monograph/item',
+        },
+      ],
+      publisherLiteral: ['Cambridge University Press,'],
+      items: [
+        {
+          '@id': 'res:i21106088',
+          uri: 'i21106088',
+          identifier: ['urn:SierraNypl:21106088'],
+          idNyplSourceId: {
+            '@type': 'SierraNypl',
+            '@value': 21106088,
+          },
+        },
+      ],
+    },
+  },
+  {
+    result: {
+      '@type': ['nypl:Item', 'nypl:Resource'],
+      '@id': 'res:b11417539',
+      carrierType: [{ '@id': 'carriertypes:nc', prefLabel: 'volume' }],
+      createdString: ['1991'],
+      createdYear: 1991,
+      creatorLiteral: ['De Grazia, Margreta.'],
+      dateStartYear: 1991,
+      dateString: ['1991'],
+      dimensions: ['23 cm.'],
+      extent: ['xi, 244 p., [8] p. of plates : ill. ;'],
+      idIsbn: ['0198117787'],
+      idLccn: ['90048121'],
+      idOclc: ['22452233'],
+      identifier: [
+        {
+          '@type': 'bf:ShelfMark',
+          '@value': 'JFD 91-4064',
+        },
+        {
+          '@type': 'nypl:Bnumber',
+          '@value': '11417539',
+        },
+        {
+          '@type': 'bf:Isbn',
+          '@value': '0198117787',
+        },
+        {
+          '@type': 'bf:Lccn',
+          '@value': '90048121',
+        },
+        {
+          '@type': 'nypl:Oclc',
+          '@value': '22452233',
+        },
+      ],
+      issuance: [{ '@id': 'urn:biblevel:m', prefLabel: 'monograph/item' }],
+      items: [
+        {
+          '@id': 'res:i13183119',
+          accessMessage: [
+            {
+              '@id': 'accessMessage:1',
+              prefLabel: 'USE IN LIBRARY',
+            },
+          ],
+          catalogItemType: [
+            {
+              '@id': 'catalogItemType:55',
+              prefLabel: 'book, limited circ, MaRLI',
+            },
+          ],
+          holdingLocation: [
+            {
+              '@id': 'loc:mal92',
+              prefLabel: 'Schwarzman Building M2 - General Research Room 315',
+            },
+          ],
+          idBarcode: [
+            '33433040712188',
+          ],
+          identifier: [
+            {
+              '@type': 'bf:ShelfMark',
+              '@value': 'JFD 91-4064',
+            },
+            {
+              '@type': 'bf:Barcode',
+              '@value': '33433040712188',
+            },
+          ],
+          requestable: [
+            false,
+          ],
+          shelfMark: [
+            'JFD 91-4064',
+          ],
+          status: [
+            {
+              '@id': 'status:a',
+              prefLabel: 'Available ',
+            },
+          ],
+          uri: 'i13183119',
+          idNyplSourceId: {
+            '@type': 'SierraNypl',
+            '@value': '13183119',
+          },
+        },
+      ],
+      language: [{ '@id': 'lang:eng', prefLabel: 'English' }],
+      lccClassification: ['PR3071 .D4 1991'],
+      materialType: [{ '@id': 'resourcetypes:txt', prefLabel: 'Text' }],
+      mediaType: [{ '@id': 'mediatypes:n', prefLabel: 'unmediated' }],
+      note: [
+        {
+          noteType: 'Bibliography',
+          prefLabel: 'Includes bibliographical references (p. [227]-238) and index.',
+          '@type': 'bf:Note',
+        },
+      ],
+      numAvailable: 1,
+      numItems: 1,
+      placeOfPublication: ['Oxford [England] : New York :'],
+      publicationStatement: ['Oxford [England] : Clarendon Press ; New York : Oxford University Press, 1991.'],
+      publisherLiteral: ['Clarendon Press ; Oxford University Press,'],
+      shelfMark: ['JFD 91-4064'],
+      subjectLiteral:
+       ['Editing -- History -- 18th century.',
+         'Malone, Edmond, 1741-1812.',
+         'Shakespeare, William, 1564-1616 -- Criticism, Textual.',
+         'Shakespeare, William, 1564-1616.'],
+      title: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus'],
+      titleDisplay: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus / Margreta de Grazia.'],
+      type: ['nypl:Item'],
+      updatedAt: 1524665261653,
+      uri: 'b11417539',
+      suppressed: false,
+    },
+  },
+];
+
+export default resultsBibs;
